[BackEnd] Added API calls to create game, add user and start game
createGame adds a new game to the games collection and sets up its
initial state as long as the inputted loginCode does not already exist.

addUser adds a player to a pre-existing but not started game as long
as the username is not taken already.

startGame sets the game state to started and calculates the targets
of all the players by shuffling the list of players and generating a
circular linked list.